+++ src/engine/Graphics/GraphicsSystem.ts
@@ -12,6 +12,7 @@ import { Particle } from '../Particles';
 import { ParallaxComponent } from './ParallaxComponent';
 import { CoordPlane } from '../Math/coord-plane';
 import { BodyComponent } from '../Collision/BodyComponent';
+import { FontCache } from './FontCache';
 
 export class GraphicsSystem extends System<TransformComponent | GraphicsComponent> {
   public readonly types = ['ex.transform', 'ex.graphics'] as const;
@@ -67,6 +68,7 @@ export class GraphicsSystem extends System<TransformComponent | GraphicsComponen
   public update(_entities: Entity[], delta: number): void {
     this._token++;
     let graphics: GraphicsComponent;
+    FontCache.checkAndClearCache();
 
     // This is a performance enhancement, most things are in world space
     // so if we can only do this once saves a ton of transform updates
@@ -94,6 +96,9 @@ export class GraphicsSystem extends System<TransformComponent | GraphicsComponen
       }
 
       this._graphicsContext.save();
+      if (transform.coordPlane === CoordPlane.Screen) {
+        this._graphicsContext.translate(this._engine.screen.contentArea.left, this._engine.screen.contentArea.top);
+      }
 
       // Tick any graphics state (but only once) for animations and graphics groups
       graphics.update(delta, this._token);
@@ -119,7 +124,7 @@ export class GraphicsSystem extends System<TransformComponent | GraphicsComponen
 
       // TODO remove this hack on the particle redo
       const particleOpacity = (entity instanceof Particle) ? entity.opacity : 1;
-      this._graphicsContext.opacity = graphics.opacity * particleOpacity;
+      this._graphicsContext.opacity *= graphics.opacity * particleOpacity;
 
       // Draw the graphics component
       this._drawGraphicsComponent(graphics);